trabalho com bugs incompleto
