+++ pasta-teste/módulo3/projeto-labex/projeto-labex/bash.exe.stackdump
@@ -1,0 +1,3 @@
+Stack trace:
+Frame        Function    Args
+000FFFF0BA0  00180062D87 (000FFFF0DA8, 00000000002, 00000000002, 000FFFFDE50)
