+++ pasta-teste/módulo3/projeto-labex/projeto-labex/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Router } from './Components/pages/Routes/Router';
+import { Router } from './Routes/Router';
