+++ pasta-teste/módulo3/projeto-labex/projeto-labex/src/Components/pages/CreateTripPage/Styled.js
@@ -1,12 +1,8 @@
-import React from 'react'
-import { useNavigate } from 'react-router-dom'
-import { goToAdminHome, goToLogin } from './Routes/Coordinator'
-import { useEffect } from 'react';
 import styled from 'styled-components';
-import Viagem from './Imagens/Viagem.jpg';
+import Viagem from '../Imagens/Viagem.jpg';
 
 
-const HomeBody = styled.body`
+export const HomeBody = styled.body`
   max-width: 100%;
   max-height:100%;
   height:730px;
@@ -18,7 +14,7 @@ const HomeBody = styled.body`
   `
 
 
-const MainHome = styled.main`
+export const MainHome = styled.main`
    display:inline-flex;
    background: inherit;
    width: 700px;
@@ -33,14 +29,14 @@ const MainHome = styled.main`
   
   `
 
-const CardHome = styled.div`
+export const CardHome = styled.div`
   display:flex;
   justify-content: flex-end ;
   align-items:center;
   padding-top:80px;
   
   `
-const ButtonStyled = styled.button`
+export const ButtonStyled = styled.button`
   margin:10px;
   background-color:#043959;
   color:#c5d9ed;
@@ -66,15 +62,15 @@ const ButtonStyled = styled.button`
  
   }
   `
-const Title = styled.h1`
+export const Title = styled.h1`
   color:silver;
   font-family: 'Chakra Petch', sans-serif;
   font-size:70px;
-  margin-top:30px;
+  margin-top:0px;
   text-shadow: 0.1em 0.1em 0.2em black;
   
   `
-const ImputStyled = styled.input`
+export const ImputStyled = styled.input`
   display:flex;
   width:600px;
   margin:20px 5px 0px 10px;
@@ -86,7 +82,7 @@ const ImputStyled = styled.input`
   font-size: 20px;
 
 `
-const HeaderStyled = styled.header`
+export const HeaderStyled = styled.header`
 backdrop-filter: blur(12px);
 border-bottom:black 5px solid;
 display:flex;
@@ -99,59 +95,3 @@ padding-top:1%;
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-const useProtectedPage = () => {
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const token = localStorage.getItem('token');
-    if (token === null) {
-      console.log("Não está logado!");
-      navigate.push(goToLogin);
-    }
-  });
-};
-
-export const CriateTripPage = () => {
-  const navigate = useNavigate()
-  useProtectedPage();
-
-
-
-
-
-
-  return (
-    <div>
-      <HomeBody>
-        <HeaderStyled>
-          <ButtonStyled onClick={()=> goToAdminHome(navigate)}>Voltar</ButtonStyled>
-          <Title>Criar Viagem</Title>
-          <ButtonStyled>Criar</ButtonStyled>
-        </HeaderStyled>
-        <CardHome>
-          <MainHome>
-            <form>
-              <ImputStyled />
-              <ImputStyled />
-              <ImputStyled />
-              <ImputStyled />
-              <ImputStyled />
-            </form>
-          </MainHome>
-        </CardHome>
-      </HomeBody>
-    </div>
-  )
-}
